refactor(components): reexported components as part of main index
+ updated import paths in components to work without aliases
+ reexported hooks as named exports
