+++ packages/components/package.json
@@ -2,7 +2,7 @@
   "name": "@deriv/components",
   "version": "1.0.0",
   "description": "deriv components",
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "private": true,
   "jest": {
     "modulePaths": [
