+++ packages/components/src/components/button-toggle/button-toggle.jsx
@@ -1,8 +1,8 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Button from 'Components/button/button.jsx';
 import HighlightWrapper from './button-highlight-wrapper.jsx';
+import Button from '../button/button.jsx';
 
 const ButtonToggle = ({ buttons_arr, className, id, is_animated, name, onChange, value, has_rounded_button }) => {
     const changeValue = selected_value => {
