+++ packages/components/src/components/calendar/calendar-button.jsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Icon from 'Components/icon';
+import Icon from '../icon';
 
 const Button = ({ children, className, is_disabled, is_hidden, icon, label, onClick }) => (
     <>
