+++ packages/components/src/components/composite-checkbox/composite-checkbox.jsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Checkbox from 'Components/checkbox/checkbox.jsx';
+import Checkbox from '../checkbox/checkbox.jsx';
 
 const CompositeCheckbox = ({ name, value, onChange, className, label, id, description, ...props }) => {
     const onClickContainer = e => {
