+++ packages/components/src/components/date-of-birth-picker/date-of-birth-picker.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { toMoment } from '@deriv/shared';
-import DatePicker from 'Components/date-picker';
+import DatePicker from '../date-picker';
 
 class DateOfBirthPicker extends React.Component {
     state = {
