+++ packages/components/src/components/date-picker/date-picker.jsx
@@ -2,11 +2,11 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { addDays, daysFromTodayTo, toMoment, convertDateFormat, isMobile } from '@deriv/shared';
 
-import DesktopWrapper from 'Components/desktop-wrapper';
-import MobileWrapper from 'Components/mobile-wrapper';
 import Input from './date-picker-input.jsx';
 import Calendar from './date-picker-calendar.jsx';
 import Native from './date-picker-native.jsx';
+import MobileWrapper from '../mobile-wrapper';
+import DesktopWrapper from '../desktop-wrapper';
 
 class DatePicker extends React.PureComponent {
     datepicker = React.createRef();
