+++ packages/components/src/components/mobile-drawer/mobile-drawer.jsx
@@ -2,13 +2,13 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 import Drawer from 'react-drag-drawer';
-import Icon from 'Components/icon/icon.jsx';
 import Body from './mobile-drawer-body.jsx';
 import Footer from './mobile-drawer-footer.jsx';
 import SubHeader from './mobile-drawer-subheader.jsx';
 import Item from './mobile-drawer-item.jsx';
 import SubMenu from './mobile-drawer-submenu.jsx';
 import SubMenuSection from './mobile-drawer-submenu-section.jsx';
+import Icon from '../icon/icon.jsx';
 
 const MobileDrawer = ({ className, id, height, width, alignment, is_open, title, toggle, children }) => (
     <Drawer
