+++ packages/components/src/components/numpad/step-input.jsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { isCryptocurrency } from '@deriv/shared';
-import Input from 'Components/input/input.jsx';
-import Button from 'Components/button/button.jsx';
+import Input from '../input/input.jsx';
+import Button from '../button/button.jsx';
 
 const getDecimals = val => {
     const array_value = typeof val === 'string' ? val.split('.') : val.toString().split('.');
