+++ packages/components/src/components/vertical-tab/vertical-tab-header-group.jsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames';
 import React from 'react';
-import Icon from 'Components/icon/icon.jsx';
+import Icon from '../icon/icon.jsx';
 
 const HeaderIcon = ({ icon, is_active }) => (
     <Icon
