+++ packages/components/src/components/vertical-tab/vertical-tab-headers.jsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 import React from 'react';
-import VerticalTabWrapper from 'Components/vertical-tab/vertical-tab-wrapper.jsx';
-import VerticalTabHeader from 'Components/vertical-tab/vertical-tab-header.jsx';
-import VerticalTabHeaderGroup from 'Components/vertical-tab/vertical-tab-header-group.jsx';
-import VerticalTabHeaderTitle from 'Components/vertical-tab/vertical-tab-header-title.jsx';
+import VerticalTabWrapper from './vertical-tab-wrapper.jsx';
+import VerticalTabHeader from './vertical-tab-header.jsx';
+import VerticalTabHeaderGroup from './vertical-tab-header-group.jsx';
+import VerticalTabHeaderTitle from './vertical-tab-header-title.jsx';
 
 class VerticalTabHeaders extends React.PureComponent {
     ref = React.createRef();
